+++ src/test/java/com/dreamypatisiel/devdevdev/domain/service/techArticle/MemberTechArticleServiceTest.java
@@ -318,55 +318,6 @@ class MemberTechArticleServiceTest extends ElasticsearchSupportTest {
                 .containsExactly(id, status);
     }
 
-    @Test
-    @DisplayName("회원이 커서 방식으로 기술블로그 북마크 목록을 조회하여 응답을 생성한다.")
-    void getBookmarkedTechArticles() {
-        // given
-        Pageable pageable = PageRequest.of(0, 1);
-
-        SocialMemberDto socialMemberDto = createSocialDto(userId, name, nickname, password, email, socialType, role);
-        Member member = Member.createMemberBy(socialMemberDto);
-        memberRepository.save(member);
-
-        UserPrincipal userPrincipal = UserPrincipal.createByMember(member);
-        SecurityContext context = SecurityContextHolder.getContext();
-        context.setAuthentication(new OAuth2AuthenticationToken(userPrincipal, userPrincipal.getAuthorities(),
-                userPrincipal.getSocialType().name()));
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        Bookmark bookmark = createBookmark(member, firstTechArticle, true);
-        bookmarkRepository.save(bookmark);
-
-        // when
-        Slice<TechArticleMainResponse> findTechArticles = memberTechArticleService.getBookmarkedTechArticles(pageable,
-                null, null, authentication);
-
-        // then
-        assertThat(findTechArticles)
-                .hasSize(pageable.getPageSize())
-                .extracting(TechArticleMainResponse::getIsBookmarked)
-                .contains(true);
-    }
-
-    @Test
-    @DisplayName("커서 방식으로 기술블로그 북마크 목록을 조회할 때 회원이 없으면 예외가 발생한다.")
-    void getBookmarkedTechArticlesNotFoundMemberException() {
-        // given
-        Pageable pageable = PageRequest.of(0, 10);
-
-        UserPrincipal userPrincipal = UserPrincipal.createByEmailAndRoleAndSocialType(email, role, socialType);
-        SecurityContext context = SecurityContextHolder.getContext();
-        context.setAuthentication(new OAuth2AuthenticationToken(userPrincipal, userPrincipal.getAuthorities(),
-                userPrincipal.getSocialType().name()));
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        // when // then
-        assertThatThrownBy(
-                () -> memberTechArticleService.getBookmarkedTechArticles(pageable, null, null, authentication))
-                .isInstanceOf(MemberException.class)
-                .hasMessage(MemberException.INVALID_MEMBER_NOT_FOUND_MESSAGE);
-    }
-
     private SocialMemberDto createSocialDto(String userId, String name, String nickName, String password, String email,
                                             String socialType, String role) {
         return SocialMemberDto.builder()
